requirements.txt
typer>=0.9.0
rich>=13.0.0
argon2-cffi>=23.1.0
httpx>=0.27.0
pytest>=8.0.0
